backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
yfinance==0.2.50
pandas==2.2.3
numpy==1.26.4
httpx==0.28.1
feedparser==6.0.11
anthropic==0.40.0
python-multipart==0.0.20
pytz==2024.2
